additional/backend_2/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
flask-swagger-ui
PyJWT
Werkzeug
Flask-Cors